Struct hermes_ipfs::HermesIpfs

pub struct HermesIpfs {
    pub(crate) node: Ipfs,
}

Hermes IPFS

Fields
node: Ipfs

IPFS node

Implementations

impl HermesIpfs

pub async fn start() -> Result<Self>

Start a new node.

Returns

HermesIpfs

Errors

Returns an error if the IPFS daemon fails to start.

pub async fn add_ipfs_file(&self, ipfs_file: AddIpfsFile) -> Result<IpfsPath>

Add a file to IPFS.

Parameters

file_path The file_path can be specified as a type that converts into std::path::PathBuf.

Returns

A result with IpfsPath

Errors

Returns an error if the file fails to upload.

pub async fn get_ipfs_file(&self, ipfs_path: GetIpfsFile) -> Result<Vec<u8>>

Get a file from IPFS

Parameters

ipfs_path - GetIpfsFile(String) Path used to get the file from IPFS.

Returns

A result with Vec<u8>.

Errors

Returns an error if the file fails to download.

pub async fn insert_pin(&self, cid: &Cid) -> Result<()>

Pin content to IPFS.

Parameters

cid - Cid Content identifier to be pinned.

Errors

Returns an error if pinning fails.

pub async fn is_pinned(&self, cid: &Cid) -> Result<bool>

Checks whether a given block is pinned.

Crash unsafety

Cannot currently detect partially written recursive pins. Those can happen if HermesIpfs::insert_pin is interrupted by a crash for example.

Works correctly only under no-crash situations. Workaround for hitting a crash is to re-pin any existing recursive pins.

Parameters

cid - Cid Content identifier to be pinned.

Returns

true if the block is pinned, false if not. See Crash unsafety notes for the false response.

Errors

Returns an error if checking pin fails.

pub async fn remove_pin(&self, cid: &Cid) -> Result<()>

Remove pinned content from IPFS.

Parameters

cid - Cid Content identifier to be un-pinned.

Errors

Returns an error if removing pin fails.

pub async fn stop(self)

Stop and exit the IPFS node daemon.

pub async fn identity(&self, peer_id: Option<PeerId>) -> Result<PeerId>

Returns the peer identity information. If no peer id is supplied the local node identity is used.

Parameters

peer_id - Option<PeerId>

Errors

Returns error if peer info cannot be retrieved.

pub async fn add_peer(&self, peer_id: PeerId, addr: Multiaddr) -> Result<()>

Add peer to address book.

Parameters

peer_id - PeerId

addr - Multiaddr

Errors

Returns error if unable to add peer.

pub async fn listening_addresses(&self) -> Result<Vec<Multiaddr>>

List of local listening addresses

Returns

Result<Vec<Multiaddr>>

Errors

Returns error if listening addresses cannot be retrieved.

pub async fn dht_mode(&self, mode: DhtMode) -> Result<()>

Sets DHT mode in the IPFS node.

Parameters

mode - DhtMode

Returns

Result<()>

Errors

Returns error if unable to set DHT mode

pub async fn dag_put(&self, ipld: Ipld) -> Result<Cid>

Add DAG data to IPFS.

Parameters

ipld - Ipld

Returns

Result<Cid>

Errors

Returns error if unable to add DAG content.

pub async fn dag_get<T: Into<IpfsPath>>( &self, path: T, ) -> Result<Ipld, ResolveError>

Get DAG data from IPFS.

Parameters

path - impl Into<IpfsPath>

Returns

Result<Ipld>

Errors

Returns error if unable to get DAG content.

pub async fn dht_put( &self, key: impl AsRef<[u8]>, value: impl Into<Vec<u8>>, ) -> Result<()>

Add content to DHT.

Parameters

key - impl AsRef<[u8]>

value - impl Into<Vec<u8>>

Returns

Result<()>

Errors

Returns error if unable to add content to DHT

pub async fn dht_get(&self, key: impl AsRef<[u8]>) -> Result<Vec<u8>>

Get content from DHT.

Parameters

key - impl AsRef<[u8]>

Returns

Result<Vec<u8>>

Errors

Returns error if unable to get content from DHT

pub async fn add_bootstrap(&self, address: Multiaddr) -> Result<Multiaddr>

Add address to bootstrap nodes.

Parameters

address - Multiaddr

Returns

Result<Multiaddr>

Errors

Returns error if unable to add address to bootstrap nodes

pub async fn bootstrap(&self) -> Result<()>

Bootstrap the IPFS node.

Returns

Result<()>

Errors

Returns error if unable to retrieve bootstrap the node.

pub async fn pubsub_events( &self, topic: impl Into<String>, ) -> Result<BoxStream<'static, PubsubEvent>>

Returns a stream of pubsub swarm events for a topic.

Parameters

topic - impl Into<String>

Returns

A result with BoxStream<'static, PubsubEvent>

Errors

Returns error if unable to retrieve pubsub swarm events.

pub async fn pubsub_subscribe( &self, topic: impl Into<String>, ) -> Result<SubscriptionStream>

Subscribes to a pubsub topic.

Parameters

topic - impl Into<String>

Returns

SubscriptionStream

Errors

Returns error if unable to subscribe to pubsub topic.

pub async fn pubsub_unsubscribe(&self, topic: impl Into<String>) -> Result<bool>

Unsubscribes from a pubsub topic.

Parameters

topic - impl Into<String>

Returns

bool

Errors

Returns error if unable to unsubscribe from pubsub topic.

pub async fn pubsub_publish( &self, topic: impl Into<String>, message: Vec<u8>, ) -> Result<MessageId>

Publishes a message to a pubsub topic.

Parameters

topic - impl Into<String>

message - Vec<u8>

Returns

Result<MessageId>

Errors

Returns error if unable to publish to a pubsub topic.

pub async fn ban_peer(&self, peer: PeerId) -> Result<()>

Ban peer from node.

Parameters

peer - PeerId

Returns

Result<()>

Errors

Returns error if unable to ban peer.