Crate hermes_ipfs

Hermes IPFS

Provides support for storage, and PubSub functionality.

Macros

pin_mut

libp2p re-export. Pins a value on the stack.

Structs

GetIpfsFile

Path to get the file from IPFS

HermesIpfs

Hermes IPFS

Ipfs

Server, Client, or Auto mode The facade for the Ipfs node.

IpfsPath

Enum for specifying paths in IPFS. Abstraction over Ipfs paths, which are used to target sub-trees or sub-documents on top of content addressable (Cid) trees. The most common use case is to specify a file under an unixfs tree from underneath a Cid forest.

MessageId

PubSub Message ID.

Multiaddr

Multiaddr type. Representation of a Multiaddr.

PeerId

Peer ID type. Identifier of a peer of the network.

PeerInfo

Peer Info type. Abstraction of IdentifyInfo but includes PeerId

SubscriptionStream

Stream for PubSub Topic Subscriptions. Stream of a pubsub messages. Implements [FusedStream].

Enums

AddIpfsFile

File that will be added to IPFS

DhtMode

Server, Client, or Auto mode

Ipld

IPLD Ipld

Traits

FutureExt

libp2p re-export. An extension trait for Futures that provides a variety of convenient adapters.

StreamExt

libp2p re-export. An extension trait for Streams that provides a variety of convenient combinator functions.

Type Aliases

BoxStream

libp2p re-export. An owned dynamically typed [Stream] for use in cases where you can’t statically type your result or need to add some indirection.

Cid

IPFS Content Identifier. A Cid that contains a multihash with an allocated size of 512 bits.

IpfsBuilder

Builder type for IPFS Node configuration.